old-tests/submission/Microsoft/selection/deleteFromDocument.htm

<!DOCTYPE HTML>
<html>
    <head>
        <title id="desc">HTML5 Selection: Delete some text from the document while it is part of a selection</title>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script src="common.js"></script>
        <script type="text/javascript">
            function RunTest()
            {
                var selection = window.getSelection();
                var p2 = document.getElementById("p2");
                var expectedResult = "abcdeuvwxyz";

                var range = document.createRange();
                range.setStart(p2.firstChild, 5);
                range.setEnd(p2.firstChild, 20);
                selection.addRange(range);
                selection.deleteFromDocument();
                p2.normalize();

                range.collapse(true);

                checkSelectionAttributes(range.startContainer, range.startOffset, range.endContainer, range.endOffset, true, 1);
                assert_equals(p2.firstChild.data, expectedResult)
            }
        </script>
    </head>
    <body onload="test(RunTest);">
        <p id="p1">Delete some text from the document while it is part of a selection</p>
        <p id="p2">abcdefghijklmnopqrstuvwxyz</p>
    </body>
</html>
